[#35104] added error handling for tickets and reservations

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/repository/SectorRepository.java
@@ -7,12 +7,12 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface SectorRepository extends JpaRepository<Sector, Integer> {
+public interface SectorRepository extends JpaRepository<Sector, Long> {
 
     /**
      * finds one sector by its id
      * @param id of the sector to be found
      * @return optional containing the sector if it was found
      */
-    Optional<Sector> findOneById(Integer id);
+    Optional<Sector> findOneById(Long id);
 }
